Add recipes and builder
Create first version that can create debian packages.

A new Cany::Recipe class is the parent for all recipes. The specification DSL
was extended to define which recipes should be included. For testing empty
bundler and rails recipes was defined. The deb helper recipes (used only for
dpkg) is the default recipes calling the debian debhelper to do all common
tasks.