recipe: introduces options for every recipes; fixes #5
A recipe can register an option. The option can afterwards be configured with
different key, values pairs (hash). The passed data for one option are merged.
The options are designed to allow other recipes (and later maybe the user) to
influence the recipe behavior.

Example: The bundler recipes allows to put/example environment variables into
the generated wrapper script. The rails recipe uses this to ensure that it is
run under RAILS_ENV=production.